CRUD add-ons, session types, favorite photographers

+++ src/main/java/com/example/picbooker/photographer_sessionType/PhotographerSessionTypeId.java
@@ -21,7 +21,7 @@ public class PhotographerSessionTypeId implements Serializable {
     private Long photographerId;
 
     @Column(name = "session_type_id")
-    private Integer sessionTypeId;
+    private Long sessionTypeId;
 
     @Override
     public boolean equals(Object o) {
